+++ docs/release_notes.rst
@@ -4,7 +4,8 @@ Release notes
 [unreleased]
 ------------
 
-- Add acoustic BSDF `[f5c57d4] <https://github.com/misuka-renderer/misuka/commit/f5c57d495597ad400a283f3620481aef802b5c7f>`_
+- Add Acoustic Path Integrator `[PR #3] <https://github.com/misuka-renderer/misuka/pull/3>`_
+- Add Acoustic BSDF `[f5c57d4] <https://github.com/misuka-renderer/misuka/commit/f5c57d495597ad400a283f3620481aef802b5c7f>`_
 
 
 misuka 0.0.0